Moved test constants inside of WrathmarkTestBase class
